feat: assign classes beside the roster, not in a mode screen (#204)
반 배정이 **전체 화면 모드**였다(`/operator/admin/assign`). 사이드바와 헤더가 사라지고
왼쪽에 명단 표를 **다시 그렸는데**, 그 표는 명단 탭의 것과 검색·필터·쪽·선택이 따로
놀았다. 그래서 고른 사람을 `location.state`로 넘겨야 했고, 넘겨도 화면이 통째로
갈아엎여서 **무엇이 이어졌는지 보이지 않았다.**

하는 일은 명단 탭에서 하던 것과 같다(사람을 고른다). **고르는 곳은 그대로 두고 넣을
곳만 옆에 연다** — 선택·필터·쪽이 그대로 남는다. 모드 화면 635줄과 라우트를 지운다.

**반 이동이 한 번에 되게 했다.** 서버에는 「배정」만 있고 「이동」이 없어서, 이미 반에
있는 사람에게 배정을 부르면 `409 DATA_INTEGRITY_VIOLATION`이 난다(같은 반이든 다른
반이든). 옮길 사람만 먼저 풀고(rollback) 이어서 넣는다 — 사용자는 한 번 누른다.
되돌리기가 실패하면 아무도 안 옮겨진 것이므로 그때는 「미배정」이라고 쓰지 않는다.

그 밖에 이 화면에서 눌러 보고 고친 것들:

- **초대 대기 행에도 「비활성」이 떴다.** 사유까지 받아 놓고 서버가 `409`로 거절했다 —
  못 하는 일을 끝까지 시켰다. 활성인 사람에게만 그린다
- **초대 재발송이 빈 배열을 보냈다.** 고른 사람이 전부 활성이면 대상이 0인데 그대로
  보내 `400`이 났다 — 정상 상황이 실패로 보였다. 보낼 것이 없으면 요청하지 않는다
- **CSV 미리보기 실패가 처리되지 않은 오류로 샜다.** 서버가 행 번호까지 주는데 화면은
  침묵했다. 그 문구를 그대로 보여준다
- 벌크바에 세로 여백이 없어 글자가 색 띠 모서리에 붙었다 — 두 층에 같은 최소 높이를
  줘서 **고를 때 표가 안 밀린다**(실측 0px)
